+++ Genesis/Changes.md
@@ -21,6 +21,7 @@ new mutable mouse event handlers.
 - Added new variations of .average(...) to RGB and added copies of those methods to Color
 - Added a new variation of .symmetric(...) to InsetsFactory
 - Added .toMap2D to TwoDimensional
+- Added .build(...) to XmlElement
 - Added .minAlong(Axis2D), .maxAlong(Axis2D) and overlapsWith(Bounds) to Bounds
 - Added .translated(Vector2DLike) to Drawer
 - Added Image.draw(...) method which can be used for drawing custom images